Amélioration de la gestion des messages : ajout des informations sur l'expéditeur dans les messages envoyés et reçus, mise à jour de l'affichage des actions en fonction de l'auteur, et correction de la récupération des destinataires dans le formulaire d'envoi de message.

+++ view/message_view_view.php
@@ -12,7 +12,12 @@
 
         <div class="action-buttons">
             <a href="/messagerie" class="action-button_back">Retour à la messagerie</a>
-            <a href="/messagerie/delete?id=<?= $message['id'] ?>" class="action-button_delete">Supprimer</a>
+
+            <?php if ($message['author'] === $_SESSION['user']['username']): ?>
+                <a href="/messagerie/delete?id=<?= $message['id'] ?>" class="action-button_delete">Supprimer</a>
+            <?php endif; ?>
+
+
         </div>
     </div>
